Fix: Resolve signed/unsigned comparison warning in Two Sum C++ solution
- Change loop variable from int to size_t to match vector::size() return type
- Add static_cast<int> when storing indices to maintain type safety
- Fixes CI/CD compilation error with -Werror=sign-compare

+++ problems/0001-two-sum/cpp/solution.cpp
@@ -17,12 +17,12 @@ class Solution {
   vector<int> twoSum(vector<int>& nums, int target) {
     unordered_map<int, int> seen;  // value -> index
 
-    for (int i = 0; i < nums.size(); ++i) {
+    for (size_t i = 0; i < nums.size(); ++i) {
       int complement = target - nums[i];
       if (seen.find(complement) != seen.end()) {
-        return {seen[complement], i};
+        return {seen[complement], static_cast<int>(i)};
       }
-      seen[nums[i]] = i;
+      seen[nums[i]] = static_cast<int>(i);
     }
     return {};
   }
